修改图片更新

+++ src/main/java/com/eighth/housekeeping/controller/UserServiceController.java
@@ -275,7 +275,9 @@ public class UserServiceController {
 		}
 		try {
 			file.transferTo(targetFile);
-			userService.deleteImageObj(objId, objType);
+			if(Constants.PORTRAIT.equals(objType)){
+				userService.deleteImageObj(objId, objType);
+			}
 			ImageObj imageObj = new ImageObj();
 			imageObj.setImageId(name);
 			imageObj.setImageType(objType);
